+++ .github/workflows/weekly.yml
@@ -2,7 +2,7 @@ name: Weekly
 
 on:
   schedule:
-    - cron: '0 0 * * 0'
+    - cron: '0 0 * * 1'
   workflow_dispatch:
 
 jobs:
