+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # 9.7.0
 - PPI-1000 - Fixes an issue, where PayLater is shown in cases where it should not be available
 - PPI-1044 - Improved compatibility of Vaulting with Store API usage and Headless setups
+- PPI-1070 - Improved locale fallback behaviour
 - PPI-1073 - Changed logging behaviour of errors happening during checkout
 - PPI-1076 - Improved information passing in Express Checkout API calls
 - PPI-1077 - Improved the availability for Google Pay and Apple Pay
-- PPI-1070 - Improved locale fallback behaviour
 
 # 9.6.5
 - PPI-1025 - Improves the performance of the installment banner in the Storefront
