--- src/Util/PaypalLocales.php
@@ -1,1503 +1,0 @@
-<?php declare(strict_types=1);
-/*
- * (c) shopware AG <info@shopware.com>
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Swag\PayPal\Util;
-
-use Shopware\Core\Framework\Log\Package;
-
-/**
- * @url https://developer.paypal.com/reference/locale-codes/
- */
-#[Package('checkout')]
-final class PaypalLocales
-{
-    /* Albania */
-    public const AL = [
-        0 => 'en_US',
-    ];
-
-    /* Algeria */
-    public const DZ = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Andorra */
-    public const AD = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Angola */
-    public const AO = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Anguilla */
-    public const AI = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Antigua & Barbuda */
-    public const AG = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Argentina */
-    public const AR = [
-        0 => 'es_XC',
-        1 => 'en_US',
-    ];
-
-    /* Armenia */
-    public const AM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Aruba */
-    public const AW = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Australia */
-    public const AU = [
-        0 => 'en_AU',
-    ];
-
-    /* Austria */
-    public const AT = [
-        0 => 'de_DE',
-        1 => 'en_US',
-    ];
-
-    /* Azerbaijan */
-    public const AZ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Bahamas */
-    public const BS = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Bahrain */
-    public const BH = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Barbados */
-    public const BB = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Belarus */
-    public const BY = [
-        0 => 'en_US',
-    ];
-
-    /* Belgium */
-    public const BE = [
-        0 => 'en_US',
-        1 => 'nl_NL',
-        2 => 'fr_FR',
-    ];
-
-    /* Belize */
-    public const BZ = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Benin */
-    public const BJ = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Bermuda */
-    public const BM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Bhutan */
-    public const BT = [
-        0 => 'en_US',
-    ];
-
-    /* Bolivia */
-    public const BO = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Bosnia & Herzegovina */
-    public const BA = [
-        0 => 'en_US',
-    ];
-
-    /* Botswana */
-    public const BW = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Brazil */
-    public const BR = [
-        0 => 'pt_BR',
-        1 => 'en_US',
-    ];
-
-    /* British Virgin Islands */
-    public const VG = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Brunei */
-    public const BN = [
-        0 => 'en_US',
-    ];
-
-    /* Bulgaria */
-    public const BG = [
-        0 => 'en_US',
-    ];
-
-    /* Burkina Faso */
-    public const BF = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Burundi */
-    public const BI = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Cambodia */
-    public const KH = [
-        0 => 'en_US',
-    ];
-
-    /* Cameroon */
-    public const CM = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-    ];
-
-    /* Canada */
-    public const CA = [
-        0 => 'en_US',
-        1 => 'fr_CA',
-    ];
-
-    /* Cape Verde */
-    public const CV = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Cayman Islands */
-    public const KY = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Chad */
-    public const TD = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Chile */
-    public const CL = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* China */
-    public const CN = [
-        0 => 'zh_CN',
-    ];
-
-    /* China Worldwide */
-    public const C2 = [
-        0 => 'zh_XC',
-        1 => 'en_US',
-    ];
-
-    /* Colombia */
-    public const CO = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Comoros */
-    public const KM = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Congo - Brazzaville */
-    public const CG = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Congo - Kinshasa */
-    public const CD = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Cook Islands */
-    public const CK = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Costa Rica */
-    public const CR = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* CÔte D’ivoire */
-    public const CI = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-    ];
-
-    /* Croatia */
-    public const HR = [
-        0 => 'en_US',
-    ];
-
-    /* Cyprus */
-    public const CY = [
-        0 => 'en_US',
-    ];
-
-    /* Czech Republic */
-    public const CZ = [
-        0 => 'cs_CZ',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Denmark */
-    public const DK = [
-        0 => 'da_DK',
-        1 => 'en_US',
-    ];
-
-    /* Djibouti */
-    public const DJ = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Dominica */
-    public const DM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Dominican Republic */
-    public const DO = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Ecuador */
-    public const EC = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Egypt */
-    public const EG = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* El Salvador */
-    public const SV = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Eritrea */
-    public const ER = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Estonia */
-    public const EE = [
-        0 => 'en_US',
-        1 => 'ru_RU',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Ethiopia */
-    public const ET = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Falkland Islands */
-    public const FK = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Faroe Islands */
-    public const FO = [
-        0 => 'da_DK',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Fiji */
-    public const FJ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Finland */
-    public const FI = [
-        0 => 'fi_FI',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* France */
-    public const FR = [
-        0 => 'fr_FR',
-        1 => 'en_US',
-    ];
-
-    /* French Guiana */
-    public const GF = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* French Polynesia */
-    public const PF = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Gabon */
-    public const GA = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Gambia */
-    public const GM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Georgia */
-    public const GE = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Germany */
-    public const DE = [
-        0 => 'de_DE',
-        1 => 'en_US',
-    ];
-
-    /* Gibraltar */
-    public const GI = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Greece */
-    public const GR = [
-        0 => 'el_GR',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Greenland */
-    public const GL = [
-        0 => 'da_DK',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Grenada */
-    public const GD = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Guadeloupe */
-    public const GP = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Guatemala */
-    public const GT = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Guinea */
-    public const GN = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Guinea-bissau */
-    public const GW = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Guyana */
-    public const GY = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Honduras */
-    public const HN = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Hong Kong Sar China */
-    public const HK = [
-        0 => 'en_GB',
-        1 => 'zh_HK',
-    ];
-
-    /* Hungary */
-    public const HU = [
-        0 => 'hu_HU',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Iceland */
-    public const IS = [
-        0 => 'en_US',
-    ];
-
-    /* India */
-    public const IN = [
-        0 => 'en_IN',
-    ];
-
-    /* Indonesia */
-    public const ID = [
-        0 => 'id_ID',
-        1 => 'en_US',
-    ];
-
-    /* Ireland */
-    public const IE = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Israel */
-    public const IL = [
-        0 => 'he_IL',
-        1 => 'en_US',
-    ];
-
-    /* Italy */
-    public const IT = [
-        0 => 'it_IT',
-        1 => 'en_US',
-    ];
-
-    /* Jamaica */
-    public const JM = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Japan */
-    public const JP = [
-        0 => 'ja_JP',
-        1 => 'en_US',
-    ];
-
-    /* Jordan */
-    public const JO = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Kazakhstan */
-    public const KZ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Kenya */
-    public const KE = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Kiribati */
-    public const KI = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Kuwait */
-    public const KW = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Kyrgyzstan */
-    public const KG = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Laos */
-    public const LA = [
-        0 => 'en_US',
-    ];
-
-    /* Latvia */
-    public const LV = [
-        0 => 'en_US',
-        1 => 'ru_RU',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Lesotho */
-    public const LS = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Liechtenstein */
-    public const LI = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Lithuania */
-    public const LT = [
-        0 => 'en_US',
-        1 => 'ru_RU',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Luxembourg */
-    public const LU = [
-        0 => 'en_US',
-        1 => 'de_DE',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Macedonia */
-    public const MK = [
-        0 => 'en_US',
-    ];
-
-    /* Madagascar */
-    public const MG = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Malawi */
-    public const MW = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Malaysia */
-    public const MY = [
-        0 => 'en_US',
-    ];
-
-    /* Maldives */
-    public const MV = [
-        0 => 'en_US',
-    ];
-
-    /* Mali */
-    public const ML = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Malta */
-    public const MT = [
-        0 => 'en_US',
-    ];
-
-    /* Marshall Islands */
-    public const MH = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Martinique */
-    public const MQ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Mauritania */
-    public const MR = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Mauritius */
-    public const MU = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Mayotte */
-    public const YT = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Mexico */
-    public const MX = [
-        0 => 'es_XC',
-        1 => 'en_US',
-    ];
-
-    /* Micronesia */
-    public const FM = [
-        0 => 'en_US',
-    ];
-
-    /* Moldova */
-    public const MD = [
-        0 => 'en_US',
-    ];
-
-    /* Monaco */
-    public const MC = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-    ];
-
-    /* Mongolia */
-    public const MN = [
-        0 => 'en_US',
-    ];
-
-    /* Montenegro */
-    public const ME = [
-        0 => 'en_US',
-    ];
-
-    /* Montserrat */
-    public const MS = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Morocco */
-    public const MA = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Mozambique */
-    public const MZ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Namibia */
-    public const NA = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Nauru */
-    public const NR = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Nepal */
-    public const NP = [
-        0 => 'en_US',
-    ];
-
-    /* Netherlands */
-    public const NL = [
-        0 => 'nl_NL',
-        1 => 'en_US',
-    ];
-
-    /* New Caledonia */
-    public const NC = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* New Zealand */
-    public const NZ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Nicaragua */
-    public const NI = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Niger */
-    public const NE = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Nigeria */
-    public const NG = [
-        0 => 'en_US',
-    ];
-
-    /* Niue */
-    public const NU = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Norfolk Island */
-    public const NF = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Norway */
-    public const NO = [
-        0 => 'no_NO',
-        1 => 'en_US',
-    ];
-
-    /* Oman */
-    public const OM = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Palau */
-    public const PW = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Panama */
-    public const PA = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Papua New Guinea */
-    public const PG = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Paraguay */
-    public const PY = [
-        0 => 'es_XC',
-        1 => 'en_US',
-    ];
-
-    /* Peru */
-    public const PE = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Philippines */
-    public const PH = [
-        0 => 'en_US',
-    ];
-
-    /* Pitcairn Islands */
-    public const PN = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Poland */
-    public const PL = [
-        0 => 'pl_PL',
-        1 => 'en_US',
-    ];
-
-    /* Portugal */
-    public const PT = [
-        0 => 'pt_PT',
-        1 => 'en_US',
-    ];
-
-    /* Qatar */
-    public const QA = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-        4 => 'ar_EG',
-    ];
-
-    /* RÉunion */
-    public const RE = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Romania */
-    public const RO = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Russia */
-    public const RU = [
-        0 => 'ru_RU',
-        1 => 'en_US',
-    ];
-
-    /* Rwanda */
-    public const RW = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Samoa */
-    public const WS = [
-        0 => 'en_US',
-    ];
-
-    /* San Marino */
-    public const SM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* SÃo TomÉ & PrÍncipe */
-    public const ST = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Saudi Arabia */
-    public const SA = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Senegal */
-    public const SN = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Serbia */
-    public const RS = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Seychelles */
-    public const SC = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Sierra Leone */
-    public const SL = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Singapore */
-    public const SG = [
-        0 => 'en_GB',
-    ];
-
-    /* Slovakia */
-    public const SK = [
-        0 => 'sk_SK',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Slovenia */
-    public const SI = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Solomon Islands */
-    public const SB = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Somalia */
-    public const SO = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* South Africa */
-    public const ZA = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* South Korea */
-    public const KR = [
-        0 => 'ko_KR',
-        1 => 'en_US',
-    ];
-
-    /* Spain */
-    public const ES = [
-        0 => 'es_ES',
-        1 => 'en_US',
-    ];
-
-    /* Sri Lanka */
-    public const LK = [
-        0 => 'en_US',
-    ];
-
-    /* St. Helena */
-    public const SH = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* St. Kitts & Nevis */
-    public const KN = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* St. Lucia */
-    public const LC = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* St. Pierre & Miquelon */
-    public const PM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* St. Vincent & Grenadines */
-    public const VC = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Suriname */
-    public const SR = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Svalbard & Jan Mayen */
-    public const SJ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Swaziland */
-    public const SZ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Sweden */
-    public const SE = [
-        0 => 'sv_SE',
-        1 => 'en_US',
-    ];
-
-    /* Switzerland */
-    public const CH = [
-        0 => 'de_DE',
-        1 => 'fr_FR',
-        2 => 'en_US',
-    ];
-
-    /* Taiwan */
-    public const TW = [
-        0 => 'zh_TW',
-        1 => 'en_US',
-    ];
-
-    /* Tajikistan */
-    public const TJ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Tanzania */
-    public const TZ = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Thailand */
-    public const TH = [
-        0 => 'th_TH',
-        1 => 'en_GB',
-    ];
-
-    /* Togo */
-    public const TG = [
-        0 => 'fr_XC',
-        1 => 'en_US',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Tonga */
-    public const TO = [
-        0 => 'en_US',
-    ];
-
-    /* Trinidad & Tobago */
-    public const TT = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Tunisia */
-    public const TN = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Turkmenistan */
-    public const TM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Turks & Caicos Islands */
-    public const TC = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Tuvalu */
-    public const TV = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Uganda */
-    public const UG = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Ukraine */
-    public const UA = [
-        0 => 'en_US',
-        1 => 'ru_RU',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* United Arab Emirates */
-    public const AE = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-        4 => 'ar_EG',
-    ];
-
-    /* United Kingdom */
-    public const GB = [
-        0 => 'en_GB',
-    ];
-
-    /* United States */
-    public const US = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Uruguay */
-    public const UY = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Vanuatu */
-    public const VU = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Vatican City */
-    public const VA = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Venezuela */
-    public const VE = [
-        0 => 'es_XC',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Vietnam */
-    public const VN = [
-        0 => 'en_US',
-    ];
-
-    /* Wallis & Futuna */
-    public const WF = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Yemen */
-    public const YE = [
-        0 => 'ar_EG',
-        1 => 'en_US',
-        2 => 'fr_XC',
-        3 => 'es_XC',
-        4 => 'zh_XC',
-    ];
-
-    /* Zambia */
-    public const ZM = [
-        0 => 'en_US',
-        1 => 'fr_XC',
-        2 => 'es_XC',
-        3 => 'zh_XC',
-    ];
-
-    /* Zimbabwe */
-    public const ZW = [
-        0 => 'en_US',
-    ];
-}
